+++ inst/resources/js/chrome_print.js
@@ -1,4 +1,23 @@
-(() => {
+{
+  let RevealReady = new Promise((resolve) => {
+    window.addEventListener(
+      'DOMContentLoaded',
+      () => {
+        if (window.Reveal) {
+          if (!window.location.search.match( /print-pdf/gi )) {
+            window.location.search = 'print-pdf';
+            return;
+          }
+          Reveal.addEventListener('ready', resolve);
+          if (Reveal.isReady()) resolve();
+        } else {
+          resolve();
+        }
+      },
+      {capture: true, once: true}
+    );
+  });
+
   let HTMLWidgetsReady = new Promise((resolve) => {
     window.addEventListener(
       'DOMContentLoaded',
@@ -40,5 +59,10 @@
     );
   });
 
-  window.pagedownReady = Promise.all([MathJaxReady, HTMLWidgetsReady, document.fonts.ready]);
-})();
+  window.pagedownReady = Promise.all([
+    RevealReady,
+    MathJaxReady,
+    HTMLWidgetsReady,
+    document.fonts.ready
+  ]);
+}